Changed the page path
Was placed on the userinfo page by mistake.
